made changes to the prompt to return better prompt and changed pricing too

+++ database/migrations/2024_02_16_151520_create_plans_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('plans', function (Blueprint $table) {
             $table->id();
             $table -> integer("planType");
-            $table -> integer("dollars");
-            $table -> integer("shillings");
+            $table -> double("dollars");
+            $table -> double("shillings");
             $table->timestamps();
         });
     }
